pyproject.toml
[project]
name = "seedboxsync-front"
version = "0.0.1a1"
description = "The frontend for SeedboxSync."
dependencies = [
  "flask",
  "pyyaml",
  "peewee",
  "seedboxsync>=3.2.0",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"